Conflicts:
	src/battle/BattleMap.java
	src/battle/entity/Unit.java

+++ src/battle/entity/Unit.java
@@ -152,7 +152,7 @@ public abstract class Unit {
             ret = (int) FastMath.sign(fractal.x);
             System.out.println(pos + "ret: " + ret);
             fractal.x = 0;
-            next = map.dijkstra((int) pos.x, (int) pos.y, dest.x, dest.y); // map.disjk : gives next position on the path
+            next = map.dijkstra((int) pos.x, (int) pos.y, (int)dest.x,(int) dest.y); // map.disjk : gives next position on the path
             geometry.setLocalTranslation(pos.x + fractal.x, Y_LEVEL, pos.y + fractal.y);
             return ret;
         }
@@ -162,7 +162,7 @@ public abstract class Unit {
             pos.y += FastMath.sign(fractal.y);
             System.out.println(pos + " ret: " + ret);
             fractal.y = 0;
-            next = map.dijkstra((int) pos.x, (int) pos.y, dest.x, dest.y);
+            next = map.dijkstra((int) pos.x, (int) pos.y,(int) dest.x, (int)dest.y);
             geometry.setLocalTranslation(pos.x + fractal.x, Y_LEVEL, pos.y + fractal.y);
             return ret;
         }
@@ -179,7 +179,7 @@ public abstract class Unit {
     public void moveTo(int x, int y) {
         dest.x = x;
         dest.y = y;
-        Vector2f d = map.dijkstra((int) pos.x, (int) pos.y, dest.x, dest.y);
+        Vector2f d = map.dijkstra((int) pos.x, (int) pos.y,(int) dest.x,(int) dest.y);
         next.x = d.x;
         next.y = d.y;
     }
